refactor(product): Product관련 Dto를 shared 경로로 변경
