--- src/main/java/com/bugzero/rarego/boundedContext/product/domain/ProductResponseDto.java
@@ -1,8 +1,0 @@
-package com.bugzero.rarego.boundedContext.product.domain;
-
-public record ProductResponseDto (
-	long productId,
-	long auctionId,
-	InspectionStatus inspectionStatus
-) {
-}
