+++ src/main/java/com/bugzero/rarego/shared/product/dto/ProductRequestDto.java
@@ -1,6 +1,8 @@
-package com.bugzero.rarego.boundedContext.product.domain;
+package com.bugzero.rarego.shared.product.dto;
 
 import java.util.List;
+
+import com.bugzero.rarego.boundedContext.product.domain.Category;
 
 public record ProductRequestDto(
 	String name,
