+++ src/main/java/com/bugzero/rarego/shared/product/dto/ProductResponseDto.java
@@ -1,0 +1,10 @@
+package com.bugzero.rarego.shared.product.dto;
+
+import com.bugzero.rarego.boundedContext.product.domain.InspectionStatus;
+
+public record ProductResponseDto (
+	long productId,
+	long auctionId,
+	InspectionStatus inspectionStatus
+) {
+}
